refactor class methods
