+++ app/stategies/BackstageStrategy.ts
@@ -1,4 +1,5 @@
-import { IStrategy } from "../context";
+import { IStrategy } from "../interfaces/strategy-interface";
+import { LAST_DATE, INCREASE, DECREASE, ZERO_QUALITY } from "../constants";
 
 export class BackStage implements IStrategy {
   sellIn: number;
@@ -10,24 +11,24 @@ export class BackStage implements IStrategy {
   }
 
   updateQuality() {
-    if (this.sellIn <= 0) {
-      return 0;
+    if (this.sellIn <= LAST_DATE) {
+      return ZERO_QUALITY;
     }
 
     if (this.sellIn <= 6) {
-      return this.quality + 3;
+      return this.quality + INCREASE.BY_THREE;
     }
 
     if (this.sellIn <= 11) {
-      return this.quality + 2;
+      return this.quality + INCREASE.BY_TWO;
     }
 
     if (this.sellIn > 11) {
-      return this.quality + 1;
+      return this.quality + INCREASE.BY_ONE;
     }
   }
 
   updateSellIn() {
-    return this.sellIn - 1;
+    return this.sellIn - DECREASE.BY_ONE;
   }
 }
